# Use `memcmp` to improve `ForEachStore`'s performance