--- a/Sources/ComposableArchitecture/SwiftUI/ForEachStore.swift
+++ b/Sources/ComposableArchitecture/SwiftUI/ForEachStore.swift
@@ -94,7 +94,10 @@ public struct ForEachStore<
   {
     self.data = store.state.value
     self.content = {
-      WithViewStore(store.scope(state: { $0.ids })) { viewStore in
+      WithViewStore(
+        store.scope(state: { $0.ids }),
+        removeDuplicates: areOrderedSetsDuplicates
+      ) { viewStore in
         ForEach(viewStore.state, id: \.self) { id -> EachContent in
           // NB: We cache elements here to avoid a potential crash where SwiftUI may re-evaluate
           //     views for elements no longer in the collection.
@@ -118,4 +121,13 @@ public struct ForEachStore<
   public var body: some View {
     self.content()
   }
+}
+
+func areOrderedSetsDuplicates<ID: Hashable>(lhs: OrderedSet<ID>, rhs: OrderedSet<ID>) -> Bool {
+  var lhs = lhs
+  var rhs = rhs
+  if memcmp(&lhs, &rhs, MemoryLayout<OrderedSet<ID>>.size) == 0 {
+    return true
+  }
+  return lhs == rhs
 }