@jshier 2022-08-30
Would this be something for `swift-collections` as well?