@tgrapperon 2022-08-30
This is possible, but it maybe has some very intricate side-effects in the general picture that I'm not prepared to delve into, especially if I would need to discuss it in Swift forums[^1]. I've asked Karoy Lorentey [on Twitter](https://twitter.com/lorentey/status/1564339015238553602) about the change, but the tweet format was maybe too short to detail the use case and implementation[^2]. According to his response, this check should already be performed. However, the two last benchmarks and [the code](https://github.com/apple/swift-collections/blob/main/Sources/OrderedCollections/OrderedSet/OrderedSet%2BEquatable.swift) suggest that this isn't the case.

Because we're falling back to `==` in case of discrepancy, the only case where it would be problematic for us would be if it would return `0` (hence `true`) with the two values containing different elements. Because `OrderedSet` has a value semantic, this is not possible, as it would mean that the same private storage would be containing different elements at the same time.

[^1]: For example, if implemented from the library, one has access to the storage, and one can rely on direct pointer comparison instead of `memcmp`. It's also possible that this can be applied to other types in `swift-collection`, etc.

[^2]: I understand that you can't start to randomly recommend low-level tricks on Twitter when you're in his position. It was already very kind from him to respond.